23 jan push

+++ recipes/models.py
@@ -15,7 +15,7 @@ class Product(models.Model):
     currency = models.CharField(max_length=3, null=True)
     productPrice = models.DecimalField(max_digits=10, decimal_places=2)
     client = models.CharField(max_length=100, null=True)
-    weight = models.DecimalField(max_digits=5, decimal_places=2, null=True)
+    weight = models.PositiveIntegerField(null=True)
     noOfSlices = models.PositiveIntegerField(null=True)
     thickness = models.PositiveIntegerField(null=True)
     recipe = models.ForeignKey(Recipe, on_delete=models.CASCADE, related_name="products", null=True)
